Rename "Entity type" interfaces to "Entity records" since that's what they actually model
