+++ packages/core-data/src/entity-types/records/menu-location.ts
@@ -1,12 +1,12 @@
 /**
  * Internal dependencies
  */
-import { Context, OmitNevers } from './helpers';
+import type { Context, OmitNevers } from '../helpers';
 
-import { BaseEntityTypes as _BaseEntityTypes } from './base-entity-types';
+import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
 
-declare module './base-entity-types' {
-	export namespace BaseEntityTypes {
+declare module './base-entity-records' {
+	export namespace BaseEntityRecords {
 		export interface MenuLocation< C extends Context > {
 			/**
 			 * The name of the menu location.
@@ -25,5 +25,5 @@ declare module './base-entity-types' {
 }
 
 export type MenuLocation< C extends Context > = OmitNevers<
-	_BaseEntityTypes.MenuLocation< C >
+	_BaseEntityRecords.MenuLocation< C >
 >;
