+++ packages/core-data/src/entity-types/records/nav-menu-item.ts
@@ -1,9 +1,14 @@
 /**
  * Internal dependencies
  */
-import { RenderedText, Context, ContextualField, OmitNevers } from './helpers';
+import type {
+	RenderedText,
+	Context,
+	ContextualField,
+	OmitNevers,
+} from '../helpers';
 
-import { BaseEntityTypes as _BaseEntityTypes } from './base-entity-types';
+import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
 
 export type NavMenuItemType =
 	| 'taxonomy'
@@ -18,8 +23,8 @@ export type NavMenuItemStatus =
 	| 'private';
 export type Target = '_blank' | '';
 
-declare module './base-entity-types' {
-	export namespace BaseEntityTypes {
+declare module './base-entity-records' {
+	export namespace BaseEntityRecords {
 		export interface NavMenuItem< C extends Context > {
 			/**
 			 * The title for the object.
@@ -102,5 +107,5 @@ declare module './base-entity-types' {
 }
 
 export type NavMenuItem< C extends Context > = OmitNevers<
-	_BaseEntityTypes.NavMenuItem< C >
+	_BaseEntityRecords.NavMenuItem< C >
 >;
