+++ packages/core-data/src/entity-types/records/nav-menu.ts
@@ -1,12 +1,12 @@
 /**
  * Internal dependencies
  */
-import { Context, ContextualField, OmitNevers } from './helpers';
+import type { Context, ContextualField, OmitNevers } from '../helpers';
 
-import { BaseEntityTypes as _BaseEntityTypes } from './base-entity-types';
+import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
 
-declare module './base-entity-types' {
-	export namespace BaseEntityTypes {
+declare module './base-entity-records' {
+	export namespace BaseEntityRecords {
 		export interface NavMenu< C extends Context > {
 			/**
 			 * Unique identifier for the term.
@@ -49,5 +49,5 @@ declare module './base-entity-types' {
 }
 
 export type NavMenu< C extends Context > = OmitNevers<
-	_BaseEntityTypes.NavMenu< C >
+	_BaseEntityRecords.NavMenu< C >
 >;
