+++ packages/core-data/src/entity-types/records/navigation-area.ts
@@ -1,12 +1,12 @@
 /**
  * Internal dependencies
  */
-import { Context, OmitNevers } from './helpers';
+import type { Context, OmitNevers } from '../helpers';
 
-import { BaseEntityTypes as _BaseEntityTypes } from './base-entity-types';
+import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
 
-declare module './base-entity-types' {
-	export namespace BaseEntityTypes {
+declare module './base-entity-records' {
+	export namespace BaseEntityRecords {
 		export interface NavigationArea< C extends Context > {
 			/**
 			 * The name of the navigation area.
@@ -25,5 +25,5 @@ declare module './base-entity-types' {
 }
 
 export type NavigationArea< C extends Context > = OmitNevers<
-	_BaseEntityTypes.NavigationArea< C >
+	_BaseEntityRecords.NavigationArea< C >
 >;
