+++ packages/core-data/src/entity-types/records/page.ts
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import {
+import type {
 	CommentingStatus,
 	Context,
 	ContextualField,
@@ -9,12 +9,12 @@ import {
 	PostStatus,
 	RenderedText,
 	OmitNevers,
-} from './helpers';
+} from '../helpers';
 
-import { BaseEntityTypes as _BaseEntityTypes } from './base-entity-types';
+import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
 
-declare module './base-entity-types' {
-	export namespace BaseEntityTypes {
+declare module './base-entity-records' {
+	export namespace BaseEntityRecords {
 		export interface Page< C extends Context > {
 			/**
 			 * The date the post was published, in the site's timezone.
@@ -140,5 +140,5 @@ declare module './base-entity-types' {
 }
 
 export type Page< C extends Context > = OmitNevers<
-	_BaseEntityTypes.Page< C >
+	_BaseEntityRecords.Page< C >
 >;
