+++ packages/core-data/src/entity-types/records/settings.ts
@@ -1,12 +1,17 @@
 /**
  * Internal dependencies
  */
-import { CommentingStatus, Context, OmitNevers, PingStatus } from './helpers';
+import type {
+	CommentingStatus,
+	Context,
+	OmitNevers,
+	PingStatus,
+} from '../helpers';
 
-import { BaseEntityTypes as _BaseEntityTypes } from './base-entity-types';
+import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
 
-declare module './base-entity-types' {
-	export namespace BaseEntityTypes {
+declare module './base-entity-records' {
+	export namespace BaseEntityRecords {
 		export interface Settings< C extends Context > {
 			/**
 			 * What to show on the front page
@@ -89,5 +94,5 @@ declare module './base-entity-types' {
 }
 
 export type Settings< C extends Context > = OmitNevers<
-	_BaseEntityTypes.Settings< C >
+	_BaseEntityRecords.Settings< C >
 >;
