+++ packages/core-data/src/entity-types/records/theme.ts
@@ -1,12 +1,12 @@
 /**
  * Internal dependencies
  */
-import { Context, PostFormat, RenderedText, OmitNevers } from './helpers';
+import type { Context, PostFormat, RenderedText, OmitNevers } from '../helpers';
 
-import { BaseEntityTypes as _BaseEntityTypes } from './base-entity-types';
+import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
 
-declare module './base-entity-types' {
-	export namespace BaseEntityTypes {
+declare module './base-entity-records' {
+	export namespace BaseEntityRecords {
 		export interface Theme< C extends Context > {
 			/**
 			 * The theme's stylesheet. This uniquely identifies the theme.
@@ -218,5 +218,5 @@ declare module './base-entity-types' {
 }
 
 export type Theme< C extends Context > = OmitNevers<
-	_BaseEntityTypes.Theme< C >
+	_BaseEntityRecords.Theme< C >
 >;
