+++ packages/core-data/src/entity-types/records/type.ts
@@ -1,12 +1,12 @@
 /**
  * Internal dependencies
  */
-import { Context, ContextualField, OmitNevers } from './helpers';
+import type { Context, ContextualField, OmitNevers } from '../helpers';
 
-import { BaseEntityTypes as _BaseEntityTypes } from './base-entity-types';
+import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
 
-declare module './base-entity-types' {
-	export namespace BaseEntityTypes {
+declare module './base-entity-records' {
+	export namespace BaseEntityRecords {
 		export interface Type< C extends Context > {
 			/**
 			 * All capabilities used by the post type.
@@ -76,5 +76,5 @@ declare module './base-entity-types' {
 }
 
 export type Type< C extends Context > = OmitNevers<
-	_BaseEntityTypes.Type< C >
+	_BaseEntityRecords.Type< C >
 >;
