+++ packages/core-data/src/entity-types/records/user.ts
@@ -1,12 +1,17 @@
 /**
  * Internal dependencies
  */
-import { AvatarUrls, Context, ContextualField, OmitNevers } from './helpers';
+import type {
+	AvatarUrls,
+	Context,
+	ContextualField,
+	OmitNevers,
+} from '../helpers';
 
-import { BaseEntityTypes as _BaseEntityTypes } from './base-entity-types';
+import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
 
-declare module './base-entity-types' {
-	export namespace BaseEntityTypes {
+declare module './base-entity-records' {
+	export namespace BaseEntityRecords {
 		export interface User< C extends Context > {
 			/**
 			 * Unique identifier for the user.
@@ -105,5 +110,5 @@ declare module './base-entity-types' {
 }
 
 export type User< C extends Context > = OmitNevers<
-	_BaseEntityTypes.User< C >
+	_BaseEntityRecords.User< C >
 >;
